+++ src/cli/commands/snapshot.ts
@@ -1,7 +1,7 @@
 import { formatSnapshotText } from '../../utils/output.ts';
 import { serializeSnapshotResult } from '../../client-shared.ts';
 import { buildSelectionOptions, writeCommandOutput } from './shared.ts';
-import type { ClientCommandHandler } from './router.ts';
+import type { ClientCommandHandler } from './router-types.ts';
 
 export const snapshotCommand: ClientCommandHandler = async ({ flags, client }) => {
   const result = await client.capture.snapshot({
