+++ src/client-metro-companion-contract.ts
@@ -1,5 +1,3 @@
-import type { MetroBridgeScope } from './client-metro.ts';
-
 export const METRO_COMPANION_RUN_ARG = '--agent-device-run-metro-companion';
 export const METRO_COMPANION_RECONNECT_DELAY_MS = 1_000;
 export const METRO_COMPANION_LEASE_CHECK_INTERVAL_MS = 250;
@@ -14,8 +12,11 @@ export const ENV_SCOPE_TENANT_ID = 'AGENT_DEVICE_METRO_COMPANION_SCOPE_TENANT_ID
 export const ENV_SCOPE_RUN_ID = 'AGENT_DEVICE_METRO_COMPANION_SCOPE_RUN_ID';
 export const ENV_SCOPE_LEASE_ID = 'AGENT_DEVICE_METRO_COMPANION_SCOPE_LEASE_ID';
 
-export type { MetroTunnelRequestMessage as MetroCompanionRequest } from './metro.ts';
-export type { MetroBridgeScope };
+export type MetroBridgeScope = {
+  tenantId: string;
+  runId: string;
+  leaseId: string;
+};
 
 export type CompanionOptions = {
   serverBaseUrl: string;
