+++ src/commands/capture-diff-screenshot.ts
@@ -8,11 +8,11 @@ import type {
   ReservedOutputFile,
   ResolvedInputFile,
 } from '../io.ts';
-import type { AgentDeviceRuntime, CommandContext } from '../runtime.ts';
+import type { AgentDeviceRuntime, CommandContext } from '../runtime-contract.ts';
 import { AppError } from '../utils/errors.ts';
 import { compareScreenshots, type ScreenshotDiffResult } from '../utils/screenshot-diff.ts';
 import { attachCurrentOverlayMatches } from '../utils/screenshot-diff-overlay-matches.ts';
-import type { RuntimeCommand } from './index.ts';
+import type { RuntimeCommand } from './runtime-types.ts';
 import { createCommandTempFile, reserveCommandOutput, resolveCommandInput } from './io-policy.ts';
 
 export type LiveScreenshotInputRef = {
