+++ src/commands/router-orchestration.ts
@@ -1,37 +1,17 @@
-import type { CommandContext } from '../runtime.ts';
-import { AppError, type NormalizedError } from '../utils/errors.ts';
-import type { CommandRouter, CommandRouterRequest, CommandRouterResult } from './router-types.ts';
+import type { CommandContext } from '../runtime-contract.ts';
+import { AppError } from '../utils/errors.ts';
+import type {
+  BatchCommandResult,
+  BatchCommandStepResult,
+  CommandRouter,
+  CommandRouterRequest,
+} from './router-types.ts';
 
-export type BatchCommandOptions<TContext = unknown> = CommandContext & {
-  steps: readonly CommandRouterRequest<TContext>[];
-  stopOnError?: boolean;
-  maxSteps?: number;
-};
-
-export type BatchCommandStepResult =
-  | {
-      step: number;
-      command: string;
-      ok: true;
-      data: CommandRouterResult;
-      durationMs: number;
-    }
-  | {
-      step: number;
-      command: string;
-      ok: false;
-      error: NormalizedError;
-      durationMs: number;
-    };
-
-export type BatchCommandResult = {
-  kind: 'batch';
-  total: number;
-  executed: number;
-  failed: number;
-  totalDurationMs: number;
-  results: readonly BatchCommandStepResult[];
-};
+export type {
+  BatchCommandOptions,
+  BatchCommandResult,
+  BatchCommandStepResult,
+} from './router-types.ts';
 
 const ROUTER_BATCH_MAX_STEPS = 50;
