+++ src/commands/selector-read-shared.ts
@@ -1,4 +1,4 @@
-import type { AgentDeviceRuntime, CommandContext, CommandSessionRecord } from '../runtime.ts';
+import type { AgentDeviceRuntime, CommandContext, CommandSessionRecord } from '../runtime-contract.ts';
 import { AppError } from '../utils/errors.ts';
 import type { SnapshotNode, SnapshotState } from '../utils/snapshot.ts';
 import { findNodeByRef, normalizeRef } from '../utils/snapshot.ts';
